Improve the consistency of `Faraday::Error::ClientError`s raised by the `RaiseError` middleware
For consistency with the 300 and 400-600 `Faraday::Error::ClientError`s we raise in `Restforce::Middleware::RaiseError`, an HTTP 413 response should raise an error with a message in the same format ("#{http_status_code}: #{description_of_error}").

This is potentially a non-backwards-compatible change, in the unlikely event that people are doing some kind of special rescuing of this error.