Disable Hashie warnings for re-defined methods such as length
Any `Client#describe` class currently results in lots of

```
You are setting a key that conflicts with a built-in method Restforce::Mash#length defined in Hash. This can cause unexpected behavior when accessing the key via as a property. You can still access the key via the #[] method.
```

warnings, bc. of the “fields”-Array as “fields” have a “length” key.

The logging layer of Hashie was introduced in `hashie v3.5.0`. And the disable_warnings method was introduced in `hashie v3.5.2`. See https://github.com/intridea/hashie/blob/master/CHANGELOG.md