When configuring an API version, check that it's valid
Salesforce API versions are string representations of a number to
one decimal place, for example `"38.0"` or `"40.0"`.

As noted by @wmadden [in his comment](https://github.com/ejholmes/restforce/issues/160#issuecomment-256888846),
if you set an invalid API version, the Salesforce API just returns
an unhelpful 404, and we raise a `Faraday::ResourceNotFound`.
It is not terribly clear when this happens the problem is your
API version, which gets put in the URL.

This tweaks `Restforce::Configuration::Option` to allow an option
to have a validator, which is used to check values provided when
they are set. You specify a lambda which is called, and may
raise an error, much like we use for specifying defaults.

We then add a validator for the `api_version` option, comparing
the provided value to a regular expression, and raising an
`InvalidOptionError` with a helpful message if it's invalid.