Properly URL escape ids with spaces in them
The client used to call CGI.escape on ids present in URLs.
This can lead to unexpected behavior where a space in an id
is encoded as a +, but is not decoded by Salesforce leading
to duplicate objects during upsert.

This commit replaces the CGI.escape calls with
ERG::Util.url_encode calls which encode space as %20.