Fix rspec warning during test runs
When running the test suite, rspec spits out the following warning:

```
`raise_error` was called with non-proc object nil
```

This is happening whenever `raise_error` is used with a non-block subject, because then, the code is executed before the expectation was set up and therefore it may yield unexpected results.

This change wraps the subject in a `lambda` so that the warning goes away and the spec works as expected. I also refactored a related spec, because due to the improved subject set up it may now become a bit simpler.