Add create, update, destroy, and upsert for Composite API.
Adds composite and composite! methods to work with the Composite API
located here:

https://developer.salesforce.com/docs/atlas.en-us.api_rest.meta/api_rest/resources_composite_composite.htm.

With the existing batch implementation, each sub-request still counts as
an individual API call. When using the Composite API, up to 25
sub-requests will count as 1 API call.

This concern is based heavily on the existing BatchAPI concern.
Differences: You can't batch multiple sub-requests into groups of 25
because the Composite API makes use of Reference ID's to allow a
developer to reference previous sub-requests. So I've put a hard limit
of 25 sub-requests and the developer will have to do their own batching.

The GET feature of the Composite API hasn't been implemented.