Explicitly define error classes for known Salesforce errors
The Salesforce REST API can return a range of `errorCode`s
representing different kinds of errors. To make these easy to
handle in your code, we want to turn these into individual,
specific exception classes in the `Restforce::ErrorCode`
namespace that inherit from `Restforce:: ResponseError`.

Today, this behaviour is achieved by taking the error code
returned by Salesforce (e.g. `ALL_OR_NONE_OPERATION_ROLLED_BACK`),
normalising it into something that looks like a Ruby class name
(e.g. `AllOrNoneOperationRolledBack`) and then looking it up
inside the `Restforce::ErrorCode` module using `const_get` to
find the correct error class to raise.

`Restforce::ErrorCode` then uses `const_missing` to dynamically
define the error class requested - so in short, if Salesforce
returns an error, a class will be defined for it.

Why change this? A few reasons:

1. It makes it easy to accidentally write wrong code. I could
   write `rescue Restforce::ErrorCode::ErrorWithWrongSpeling`
   and this code would be valid, as `const_missing` would define
   the constant for me. But I wouldn't be rescuing what I thought
   I was. (The same goes for referring to an error that it totally
   wrong - it is not just spelling mistakes.)
2. The library doesn't have any clear list anywhere of the errors
   it *might* raise, and it is impossible to generate such a list.
   This doesn't make the library terribly easy to work with.
3. The dynamic definition of constants raises the spectre of
   exploits of users' code which might allow them to take up lots
   of the user's memory by creating many exception classes. This
   would only be an issue, of course, if you were doing some kind
   of dynamic lookup yourself.

This code changes the approach so we take a list of known error
codes from the Salesforce documentation  and define the error
classes for them upfront.

The bad part of this change is that we have to decide what to do
when we get an error back from Salesforce that isn't predefined.

I've decided to go for raising a generic `Restforce::ResponseError`
in this case, and logging a warning telling the user to report
the missing error on GitHub. __This has the potential of being
a breaking change as there could be real Salesforce errors that
we don't know about, which users of Restforce will no longer be
able to rescue (as they don't exist) and which will raise a new
generic `ResponseError`).

We will also log a warning if someone refers to an class
in the `ErrorCode` namespace that doesn't exist.

Thanks to @presidentbeef for his initial work on this in #550.