Fix Lint/MissingSuper Rubocop offence in `Restforce::Concerns::Picklists::PicklistValues`, which is an `Array`
Somewhat controversially, the `PicklistValues` class inherits from
`Array`, but it never calls `super` in the `#initialize` function,
so the array is not set up properly, which upsets Rubocop.

Today, `#initialize` stores some instance variables, works out the
elements to go in the array and then uses `Array#replace` to fill
the array. This uses `super` instead to fill it.