Adapt tests to work with Faraday 0.x and Faraday 1.x
There are some parts of Faraday's internals - which we test
to check that the package is working - which are subtly different
between Faraday 0.x and Faraday 1.x onwards.

This updates our tests to reflect this.

The error handling behaviour for authentication errors looks
wrong in Faraday 0.x, but I won't fix this for now as it will
be a breaking change for legacy Faraday users.