When a response claims to be gzipped but isn't, return the body as it is and don't explode
In the past couple of months (see #761), users have reported an
issue where some API calls raise a `Zlib::GzipFile::Error`:

```
```

The root cause [seems][1] to be that Restforce is seeing a
`Content-Encoding` response header and thinking that it needs to
decompress the body, when in fact, the body is not compressed.
When you try to decompress something that isn't compressed with
Gzip, it raises an error.

This catches that error and returns the body as is.

Fixes #761.

[1]: https://github.com/restforce/restforce/issues/761#issuecomment-1329012277